+++ packages/experiment-tag/CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.1](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-tag@0.17.0...@amplitude/experiment-tag@0.17.1) (2026-03-20)
+
+
+### Bug Fixes
+
+* stricter check for mobile mode ([#293](https://github.com/amplitude/experiment-js-client/issues/293)) ([ee87833](https://github.com/amplitude/experiment-js-client/commit/ee87833ddfc7839a2196d6706d1a5e6eb6ee51cb))
+
+
+
+
+
+# [0.17.0](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-tag@0.16.0...@amplitude/experiment-tag@0.17.0) (2026-03-19)
+
+
+### Bug Fixes
+
+* expose full dom-mutator API on ampDomMutator ([#287](https://github.com/amplitude/experiment-js-client/issues/287)) ([c1bed9f](https://github.com/amplitude/experiment-js-client/commit/c1bed9f4936ddea531bef0a44a9a80033c49aab8))
+* remove subscription dedupe for pageChangeSubscribers ([70bbd35](https://github.com/amplitude/experiment-js-client/commit/70bbd35abcef622ebd80f4b1cfe5747ed830607b))
+* Replace instanceOf checks with type casting when deduping multiple initializations ([#286](https://github.com/amplitude/experiment-js-client/issues/286)) ([b46f504](https://github.com/amplitude/experiment-js-client/commit/b46f504c758b4b841b2123fd103809ffa6842e49))
+
+
+### Features
+
+* site as shell ([#261](https://github.com/amplitude/experiment-js-client/issues/261)) ([5da84e8](https://github.com/amplitude/experiment-js-client/commit/5da84e8fc2a1cf5c4b807aab0c50e90b4e897336))
+
+
+
+
+
 # [0.16.0](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-tag@0.15.3...@amplitude/experiment-tag@0.16.0) (2026-03-11)